feat: 新增treeNode实现

+++ ssc-ddd/domain/ticket/src/test/java/indi/melon/ssc/ticket/domain/ticket/BoxManagerTest.java
@@ -23,7 +23,7 @@ class BoxManagerTest {
     private BoxManager boxManager;
     private TicketBoxRepository repository;
     private static final int cacheSize = 3;
-    private static final int currentMaxTicket = 0;
+    private static final long currentMaxTicket = 0L;
     private static final int ticketNum = 10;
 
     @BeforeEach
